Fix create_room: disambiguate Revit 2026 ElementId overloads
Two call sites in rooms.py (_resolve_level and _set_elemid_param) build
ElementIds from bare Python ints. On Revit 2026 IronPython this hits

    TypeError: Multiple targets could match:
      ElementId(BuiltInParameter), ElementId(BuiltInCategory),
      ElementId(Int64)

because Python int can box to any of the three. Cast through
System.Int64 to nail down the Int64 overload, same fix that was applied
to create_dimension in 99eef91.

Currently latent in practice: the default _resolve_level path falls
through to the lowest-elevation level (no ElementId built), and
_set_elemid_param is only reached when the optional upper_limit_id is
supplied. Any caller exercising explicit level_id or upper_limit_id
would otherwise have hit the swallowed TypeError and silently fallen
back to nearest-level resolution.

+++ revit_mcp/rooms.py
@@ -16,6 +16,7 @@ batch can be added later if a real workflow demands it.
 
 from pyrevit import routes, revit, DB
 from utils import normalize_string, element_id_value
+from System import Int64
 import json
 import traceback
 import logging
@@ -51,7 +52,7 @@ def _resolve_level(doc, level_id=None, elevation_mm=None):
     """
     if level_id is not None:
         try:
-            lid = DB.ElementId(int(level_id))
+            lid = DB.ElementId(Int64(int(level_id)))
         except Exception:
             return None, "level_id must be an integer"
         el = doc.GetElement(lid)
@@ -206,7 +207,7 @@ def _set_elemid_param(elem, bip, value_id):
     p = elem.get_Parameter(bip)
     if p is not None and not p.IsReadOnly:
         try:
-            p.Set(DB.ElementId(int(value_id)))
+            p.Set(DB.ElementId(Int64(int(value_id))))
         except Exception:
             pass
